+++ server/src/main/java/com/ryc/api/v2/applicant/service/ApplicantService.java
@@ -117,6 +117,10 @@ public class ApplicantService {
 
   @Transactional
   public void deleteApplicants(List<String> applicantIds) {
+    if (applicantIds.isEmpty()) {
+      return;
+    }
+
     eventPublisher.publishEvent(new ApplicantDeletedEvent(applicantIds));
     applicantRepository.deleteAllByIdIn(applicantIds);
     applicationRepository.deleteAllByApplicantIds(applicantIds);
