+++ server/src/main/java/com/ryc/api/v2/auth/service/AuthService.java
@@ -130,6 +130,9 @@ public class AuthService {
   @Transactional
   @EventListener
   protected void handleAdminDeletedEvent(AdminDeletedEvent event) {
+    if (!refreshTokenRepository.existsByAdminId(event.adminId())) {
+      return;
+    }
     refreshTokenRepository.deleteByAdminId(event.adminId());
   }
 }
